erpnext_mcp/requirements.txt
mcp
google-genai
pandas
openpyxl
pypdf
pdf2image
pytesseract
requests
python-dotenv
fastapi
uvicorn
python-multipart
